+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to
 
 ### Added
 
+- Pass the user_id param on check usage limits.
+  [#2387](https://github.com/OpenFn/lightning/issues/2387)
+
 ### Changed
 
 ### Fixed